requirements.txt
addict==2.4.0
aistudio-sdk==0.3.6
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
autodistill==0.1.29
autodistill-grounded-sam==0.1.2
autodistill-yolov8==0.1.4
bce-python-sdk==0.9.45
blinker==1.9.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
comm==0.2.3
common==0.1.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
fonttools==4.59.2
fsspec==2025.7.0
future==1.0.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
imagesize==1.4.1
importlib==1.0.4
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
labelImg==1.8.6
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
modelscope==1.29.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs==4.4.0
polars==1.32.3
prettytable==3.16.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
rf-groundingdino==0.2.0
rf-segment-anything==1.0
roboflow==1.2.7
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentry-sdk==2.35.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
supervision==0.26.1
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.0.81
ultralytics-thop==2.0.16
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yapf==0.43.0
